WIP: Introduce static method Link.normalize(title)
Moves normalisation and basic validation logic from
Link.__init__ and Link.parse into a new static method
which can be used on a title, without a Site, to perform
checks and fixes on the string that are not site specific.

This change results in several of the exceptions raised
in wikibase_tests to change from APIErrors (from the server)
to being Pywikibot client exceptions.  These are the same
exception changes as Ibea8f84242a6e938882059f5d5fa28394b8a7a23

TODO:
- Investigate namespace validation using new Namespace class.
- Many page_tests report the error at the bottom of __init__
  due to Page.botMayEdit looking for {{nobots}} (Bug 69384)

Bug: 61832 (fixing workaround that was merged.)
Change-Id: I7021e3d7e40d72fd74709f396e967c9803248c52