Add pywikibot.cache
Currently we have CachedRequest, but that is hardcoded to use
a filesystem cache, not letting us take advantage of faster caches
like redis or memcached if available.

pywikibot.cache is a copy of werkzeug.contrib.cache with minor changes
to not require a dependency up on werkzeug (a WSGI utility library).

This also introduces pywikibot.Cache() for bots to take advantage of
whatever cache is setup.

Change-Id: Ie8594fb7323a23a330dd67fd0571dc1a0ea150fb