WIP - API XML support
There are a few parts of the API which can only be accessed safely via
the XML API. In particular, import/export is done exclusively in XML
dumps by MediaWiki and there are no plans to replace such format ever.

The intention is to have a config variable to enable using XML for all
API calls, however it is not expected that people will use this except
when the JSON doesn't contain the information they need.  The most
important part of this patchset will be ensuring that mediawiki errors
are handled properly.

Not all of the tests pass.

The WbTime changes need to be split off.

Conversion of strings in the XML to int needs to be done where
appropriate.

Change-Id: I2e19bdb26afd009c87616539f823f51b7046ccc5