Use single dict for config disambiguation comment
The configuration item disambiguation_comment used a complex
structure of languages within families.  This required that
the default config had a non-empty dict consisting of a value
for each family, so that adding a configuration for a language
was simple.

This change detects the old disambiguation_comment and informs the
operator to consult README.config which instructs them to use the new &
simpler single level config structure of disambiguation_summary, which
uses the same syntax as replicate_replace.

The single level config structure of disambiguation_summary is supported
by i18n.translate as xdict may be a string.

Change-Id: Ie0ceedad9c56ea23a9ada18313ee80d03ddb5390