WIP: Make Family.languages_by_size dynamic
Currently languages_by_size data is obtained from wikistats.wmflabs.org
via a script which re-writes the family files for familes which have
data in the wikistats database.

The languages_by_size data is not heavily used throughout pywikibot
scripts in the repository.  This change replaces the current system with
extraction of the statistics from the API of each site in the family,
done on demand.  On a cold cache, the process takes around 8 minutes
for all Wikipedias, and 5 minutes for all Wiktionary sites.  Once
the data has been built, it is cached for 30 days and loads within
one or two seconds.

TODO: wikimedia_sites script also found new sites, not already in the known
languages list.  Functionality to achieve this has been coded as an
AutoSubdomainFamily, which needs to merge first and then the Wikimedia
families can use that as a replacement for the wikimedia_sites script.

Bug: T78396
Change-Id: I44cb513bb12e09162b4b6332cbe15c95d7e8087a