Replace Bot.getOption() with 'options' property
Bot._options is a new variable of OptionsDict type (subclass of dict
for internal use), which allows to get and set bot options with the
same 'options' property.

The standard 'update' method of the dict class has also been subclassed
in OptionsDict, and replaces the Bot.setOptions() method.

Every use of getOption() in scripts has been replaced.

Change-Id: Ia26b64bcab49a89dd907de6ba5cabb4d8ed3bb76