Use source for __init__ of Page subclasses
Page.__init__ accepts 'source' as the first and only mandatory
parameter, which may be a Site or a Link.

Category, ImagePage and User all use the same name for this
parameter.

WikibasePage, ItemPage, PropertyPage and Claim were created with 'site'
as the name of this first parameter.

Claim was a subclass of PropertyPage, but is now a subclass of
the new Property class.

This change renames the 'site' to 'source' for all subclasses of Page.
Claim and Property continue to use 'site'.

Change-Id: I92c4865340f6f3eb3efcb4de43d008da50469164