[FEAT] Add min_len to check a generators minimum length
In interwiki_graph, it's not important how many elements are in there,
just that there are at least two, so it would be a waste to query more
than two. This is used in the category script to determine if it has at
least one element.

It either returns the generator for all elements if there are enough in
it, or False otherwise.

Change-Id: I7112838745660829b6b7edaa42acaccfb86a0667