[FIX] Changing family files in user-config.py
It is not possible to change the family files in user-config.py directly
because 'register_family_file' modifies the global and not the local
family_files dictionary. After executing user-config.py all changes are
written into globals(), so the unchanged 'user scope' family_files is
different than the modified 'global' family_files which leads the the
situation that the changes are reverted.

As family_files is the only configuration variable which is modified by
a function this now ignores all changes to family_files made in the
user-config.py and warns if someone has changed it (outside the
functions).

Change-Id: I46bb039f4959644e7f6c93e82a62478b06228781