Use ordereddict for bot and pagegenerators args
bot and pagegenerators each contain a long string consisting of help
information for the user about each parameter which can be handled.

This change splits those into a dict with help for each argument,
and reconstructs the long string dynamically.

Also rewrites showHelp as show_help, which allows the indent level
to be adjusted, and multiple substitutions to occur to the
help text.

Change-Id: If191fc60ae1bd6d926f944ec0c4dfde3194cbecc