Extend DeprecationWrapper and redirect_func
DeprecationWrapper can now deprecate keys within a dict, items within
an array, or an instance method of a object instance.

redirect_func now detects the name that its returned type is given,
and supports all types of functions and methods, with the only problem
being that on Python 2 it is impossible to determine the class of a
static method, so the class needs to be given to redirect_func.

Added more methods to the wrapper class, providing a better proxy such
that epydoc is able to inspect the wrapped object.

Tests for most of the deprecation layer, excluding the module wrapper.

Change-Id: Ib8d9407726280ac137603d3eede85a33e4035d71