Generic preloading of page data
Instead of preloading only revisions, info and categoryinfo
allow callers of preloading routines to specify a dict of
page data to preload.  Implemented in:
    APISite.preloadpages
    PreloadingGenerator
    GeneratorFactory

e.g. To disable revisions data being preloaded, but adding links
     use:
         {'revisions': False, 'links': True}
     or to exclude the content and summary from the revisions data:
         {'revisions': 'ids|flags|timestamp|user'}

The available page data which can be preloaded depends on what is
implemented in api.update_page

Bug: 65161

Change-Id: If365ec24da11db08007e811a544f85943aa25a6f