[FEAT] Page: Rewrite of title
This restructures the title method of Page:
* asLink is replaced by brackets (though brackets does work a little
  different)
* allowInterwiki, forceInterwiki are replaced by interwiki

This now supports two modes of interwiki prefixes:
* valid prefixes, which could cause an exception of there is none
* fake prefixes, which are written before the brackets (if enabled)

There are seven different combinations now, depending on whether a fake
prefix is used all the time or just if required and whether an interwiki
prefix is used all the time or just if required.

It uses uses the interwiki_prefix function to get the valid prefix from
the site. It supports two level deep interwiki links, if the first
interwiki prefix goes into the family of the page. For example if the
page is in the German Wiktionary and insite is set to the English
Wikipedia it'll first search for an interwiki prefix to the Wiktionary,
if the site also supports an interwiki prefix on it's own to the German
Wiktionary.

In theory it could do a search of all prefixes, but which would cause
many requests, especially if there is no path. It's also at least as
good as the fake prefix.

By default this could cause an exception if there is no interwiki, so
all outputs have been changed to use no interwiki prefix at all and to
show the site it's working on directly or it's using a fake prefix.

The fake prefixes may not be valid (but they can be) and are like the
prefixes generated by asLink.

It introduces brackets as additional parameter because it redefines
'insite': Instead of the configured it uses the page's site by default
(which would use a self referencing interwiki prefix if forced). It's
still possible to get the configured site by using 'True'.

A dedicated 'link_label' has been added wich is a userdefined link
label. It can also hold a boolean to use the default title (e.g. for
interwiki links) or to use no label at all. If set to an empty string it
actually uses that and is not the same as False. If the link label and
the actual title are identical (it doesn't matter if the label is set
via bool or string) the label won't be used.

This also supports all parameters independelty, so it's possible to get
a textlink (colon in front of files and categories) without using
brackets. Only 'insite' is ignored if 'interwiki' is set to never.

Change-Id: Idde31761411cd4a1ac0d0e5b5b10c5b02a17f965