[FIX] MIME/Request: Urlencode MIME headers
If using MIME the header must be ASCII, but when the file name is using
non-ASCII letters those will be directly added and the server can't
interpret them. The server returns with a rather confusing message:

  missingparam: One of the parameters filekey, file, url, statuskey is
  required

This patch is encoding the parameters which aren't compatible with ASCII
with the site's encoding and then using urllib's quote_from_bytes (or
just quote in Python 2) to convert them into ASCII strings. If the
encoding is not given all entries must be ASCII strings, or the
UnicodeError occurring when encoding the string to ASCII will be
reraised.

Change-Id: Ifb0eb96051c6f4f29676ef2d5a7d4658302ddf5c