Fix claimit -exists functionality
The implementation of -exists meant options 's' and 'q' were not
operative, as it did not set any sources or qualifiers on the claim
to be added, yet it test the missing sources and qualifiers against
those already on the item.  As a consequence, 's' and 'q' would only
prevent a new claim being added when the item had no sources or
qualifiers respectively, as then both sides of the equality test
would be empty and therefore equivalent.

Also, the utility method which did the comparison did not understand
the structure of Claim.sources, and did not test equality correctly.

This change slightly re-interprets the meaning of 's' and 'q',
so that 's' means 'allow adding claims without a source' and 'allow
adding claims even if a claim on the same property has the same source'
and 'q' means 'allow adding claims even if the property has a claim
with a qualifier.'

Bug: T69284
Change-Id: I4c1c0b8b786ceb116497b192b8167cf8f63dc30b