Allow some Claim methods to be performed locally
ItemPage.editEntity can be used to submit multiple changes
to the repository at once.  To do this, the ItemPage in memory
is altered, and editEntity determines the difference between the
in memory copy and the server record.

Most of the ItemPage methods that related to Claims submit the
change immediately, which prevents these methods being used to
batch up a set of changes to the ItemPage before being submitted
to the repository together.

This change only allows ItemPage.addClaim and Claim.addSources and
Claim.addQualifier to be used to alter the in memory copy.

Bug: T57148
Change-Id: Id41ff0c6c30a64ab4add1a93a7f497e142900eca