[FEAT] Page: 'force_ns' argument in Page.__init__
If set to True this does not use any eventual namespaces from the title
but uses the given namespace. With False it always chooses the namespace
from the title and verifies that the given namespace to the constructor
is the same.

Set to None it acts like before and uses the namespace from the title if
there is a namespace which overwrites the given ns. This value should be
avoided as it can lead to strange behavior: If a user is asked to input
a category name without the namespace and the user inputs
'Template:Foobar' (because the category is 'Category:Template:Foobar')
it would fail because the page's namespace afterwards is the template
namespace. In this case force_ns should be set to 'True'.

The ns parameter can now also be set to None to allow a similar
behavior together with force_ns=False compared to the current version,
because force_ns=False and a ns != None will verify that the actual
namespace determined via the title is the same as the given namespace.

If applicable it uses the force_ns parameter (e.g. from API responses
which always add the namespace). Sometimes a script combined the
category namespace name and the title and passed that. If it's known
that the namespace is missing, force_ns=True together with ns can be
used instead (without the need to add the category namespace to the
title).

Bug: T67262
Change-Id: I4e7e67ec855d366fcf90005481093de64e5465c1