Add site_detect.load_site
load_site returns either an APISite or a NonMWAPISite.

MWSite is improved:
* option `quick` to speed up detection, exiting the
  constructor when it has found a API endpoint URL
* method `verify`, which checks that the detected
  API endpoint URL returns content similar to a
  valid API endpoint

Note that this patch originally included a dedicated
algorithm in order to detect IWM sites using `httplib2`.
As pywikibot has switched to `requests`, and a
better MW HTML parser was introduced in 2f28bd3,
the dedicated algorithm was no longer needed, and
beneficial parts of it were merged into MWSite.

Bug: T88601
Change-Id: Ifb8c5930c21e4a90b08ef958329e22a0b97a57ae