[FEAT] More flexible threaded generators
The original ThreadedGenerator supported giving a target and also
subclassing at the same time. This is deprecating ThreadedGenerator and
replaced by generators which are specialized in either a target already
being an iterator, or being a callable or having no target at all and
allow subclassing.

This doesn't fix that ThreadedGenerator supports empty lists (or
similar) but instead that script should use a specialized generator
instead.

Bug: T85667
Change-Id: I8e4a504d1673c0c9fb2e60647e432b22f7a3d48f