[FEAT] Cosmetic changes: Support special pages fix
This is a more flexible version of the 'specialpages' fix in
pywikibot/fixes.py. It fixes the case of links to special pages and
additionally changes the links to either an alias or the canonical name.
Either the first alias is always choosen or only instead of the
canonical name.

Change-Id: If98cf44fa3462838ad238a21daa42e6687974784