[FEAT] tools: redirect_method to deprecate methods
The redirect_method function wraps around redirect_func and makes it
easier to deprecate methods (functions of a class). So instead of the
template

  A = redirect_func(B, old_name='A', class_name='D')

it's reducing the duplication of A and makes sure that both are the
same:

  redirect_method(D.B, 'A', D)

As the class name is known after creation of it, this call needs to be
executed afterwards instead of inside, which is why B must be defined
absolutely. It is possible to use the string name for and avoid the
duplication of D:

  redirect_method('B', 'A', D)

Change-Id: Ic3ded7eba1844276b2ea28533ceb1e7bdcb71d19