[FEAT] Site: Dynamic linktrail support
This adds a method to get the dynamic linktrail as reported by the API. To do
that it needs to convert the PCRE regex into a Python regex. The regex returned
by the API will also match any characters after that in a second group so it
needs to return only the contents of the first group.

It also adds a full_chr function to tools which converts all codepoints into
unicode strings indepedently of the sys.maxunicode value. It uses the code from
pywikibot.page.html2unicode for that.

Change-Id: I8673565b2a9a28ced83ab589d8361696c80c1672