[WIP] [FEAT] page: Implement a RevisionCache
This implements a RevisionCache, which stores chunks of revisions and
mostly requests uncached revisions. If a certain number of revisions are
requested it requests at first the newest revisions until the most
recent cached one. When less revisions are returned than requested it'll
return as many cached revisions as requested. If not enough revisions
are cached, it'll then request revisions again but this time starting
not from the newest revision possible but from the last cached revision
until the next cached revision chunk (if there is such). It then returns
and caches those revisions until enough are returned or the next cached
revision is found. If the next cached revision is found it's starts over
again with returning cached ones and then rerequesting missing
revisions.

The caches can be shared between multiple Page instances using the same
title, although that is very likely not thread safe.

TODOs:
* Write tests (some revisions could be requested, then some could be
  deleted and it could verify that it'll fill the gaps)
* Handle deleted and undeleted revisions (it could instead of blindly
  returning the cache request just the IDs and verify it via that)

Change-Id: I04f619480049a415558e5588653da089e4903645