Adding splitwarning.py from compat to pywikibot-core.
Porting this might help in solving the warnfile.py porting issue which is a
subtask in Pywikibot-compat to core Migration. Presently, I could not
figure out the kind of WARNING which interwiki.py may generate in it's
log file and which is recognised by the splitwarning.py script for
functioning. As such the script is yet to be properly tested.

Change-Id: Icd4b512f41831e84bf0f55541d6ed052a2690acd