lonelypages: Simplify orphan configuration
Instead of storing the orphan templates separatly (one mapping for the
used template and one regex for the aliases) it's dynamically creating
both depending on the given aliases and names. It'll also automatically
choose all template namespace prefixes in the given language. The regex
to match the template will be also compiled only once and not for each
page.

Change-Id: Ie9c86c196ef33ff7bb336bb2bac9e688b41938a3