Access closed wikis and deprecate `obsolete`
Site.obsolete was a boolean set in BaseSite,
and used for a variety of meanings.

Family.interwiki_replacements and .interwiki_removals
have previously been created to provide clear meaning
to two usages of obsolete.

Now Family.code_aliases allows deprecated codes to
be replaced with new codes.

Also Wikimedia sites that had codes in obsolete
for closed wikis are now added to the list of
valid codes that can be used.

Bug: T74674
Change-Id: I6207b18b8225f7ddc1349dffe321d93a8bf25b00