[IMPROV] More sensible input()
Instead of stripping text from both sides it just strips from from the
right (as there a space is actually added). It now also checks whether
there is a color code at the end. In that case it checks the character
before that color code instead of checking the last character to insert
the default value.

Change-Id: I7447bf199a89bfe67702778203511b96ee1df530