[WIP][FEAT] pagefromfile: Simplified file structure
Instead of requiring a start and end marker it now supports just a combined
marker (called core-mode). It also supports to define the start marker
automatically by the first line of the file.

With the core-mode it also allows to improve the file format by adding the
capability to change the marker midway (making it streamable) and specifying
the page title explicitly outside the page's text.

If after the marker follows no space or new line, all text until the next space
or new line is the new marker. If after the marker (or markers) is a space the
text until the new line is the page title. The content, different to the old
compat compatible mode, starts after the new line.

Change-Id: Iff3e11c6237506805003409e13baf6f28f1fe8e8