[WIP][FEAT] textlib: Iterative template extraction
Instead of using regex and markers to extract templates this approach first
scans the text for all templates by counting the brackets basically. After the
positions of the templates are found it looks inside the template to determine
where parameters start and end.

Separating both is necessary because it's not possible to determine if opening
brackets belong to a closed template.

It has the same result as mwparserfromhell and also supports certain corner
cases the current regex variant does not.

Change-Id: I5d0bd78cf153d0948f8e52468fc882602ca0134f