[WIP]: ProofreadPage refactoring (use application/json fmt)
Loading text as application/json is more efficient and less error prone,
as no regex processing is needed unless the page is not existing yet.

ProofreadPage has been heavily refactored and raises exceptions if
correct json format is not recognised.

This approach has one issue to be solved, that is how to make sure that
all BasePage methods using site.loadrevisions are not broken.

Change-Id: I5128fbb59e0a78498057ceece095e917c1f77e86