[FEAT] Family: Allow family name as class name
This first tries to load the module attribute using the family's name (e.g.
WikipediaFamily when the family's name is wikipedia). As a fallback it uses the
attribute with the name "Family".

generate_family_file now uses the new name (and also super). Importing the
Family class as Family into a file doesn't lead to collisions anymore.

Change-Id: I5a79a56380490655eb179682eaa4132586c14029