[FEAT] Support any class name in family files
Previously the class name must be "Family" as it only used that variable of a
module. This allows to use any name although it recommends to use the name of
the family (e.g. the wikipedia family would use WikipediaFamily). It also
supports multiple families in one file (although it will warn that there are
families which don't use an appropriate name).

Change-Id: Id029ea32fdd25db723a10dd0bb57c07508a690f1