Rework Bot class: use attribute parameters
In this change, I have added a BotOption class that is used to define
options for the bot class. Currently, it basically behaves just like
a normal attribute, but in the future, I hope to add code that auto-
generates a __doc__ for the class from it.

This keeps the original API (with availableOptions and getOption)
intact; I have tested this with touch.py, which still works while making it much
easier to move other existing bots to the Bot class - they can just keep using
their normal attribute access method.

Tests in bot_tests.py check bot the new and the deprecated APIs.

Change-Id: I613d13db51e540f94d64c77908acf21cccfa2637