[FEAT] Site: Follow redirect chain
This adds the possibility to get a certain redirect along the redirect
chain. It then only throws CircularRedirect exception if the length of
the chain must be known (e.g. to get the last element) or in the default
mode which returns the immediate target but still checks for circular
redirects (to remain backwards compatibile).

Bug: T85711
Change-Id: I69880ee98b3019fc9edfd2d315ee964cb36ec6b2