[FEAT] Add basic support for signature and replace getargspec
With Python 3.3 (the most oldest version of Python 3 supported) the function
`signature` in the `inspect` module was added which replaces the `getargspec`
function in the same module. This implements a basic implementation of it to
use `signature` in both versions. Since Python 3.3 it'll use the Python
implementation of it, but for backwards compatibility return `None` on
`ValueError``.

As `inspect.getargspec` has been deprecated in Python 3, it is replaced by
`pywikibot.tools.signature`.

Bug: T106209
Change-Id: I421bd03f9fdae6257317f8971dc8070166da0592