Dynamically add appropriate tokens
This patch detects if the request is using actions or submodules which
require tokens and adds them automatically to the request. As this
feature was added in 1.24wmf19 to ParamInfo it still has a fixed mapping
for older versions of MediaWiki. Those only support actions and not that
submodules require tokens.

Bug: T78393
Change-Id: I9e93ca978f67eb7c2517bd32f2efeb664833dbb3