[FEAT] Request: Dynamically determine write
This determines dynamically whether the request is a writing request
using ParamInfo. It not only checks the action itself but also any
submodules of that action. This also applies to determining whether the
request must be POSTed and is used by the Flow extension as only certain
submodules were writing. As it wasn't supported to be determined
automatically until this patch, they specifically defined whether they
need to be POSTed which is undone with this.

The Siteinfo class must be also adjusted because it is required for
ParamInfo to work. ParamInfo requires the version number which is
returned by the Siteinfo which itself does a request which requires
ParamInfo unless it's told what to do.

Bug: T74207
Change-Id: Id9708ade5434c87d95334f1c66f827e2f408997d