api.py: buffer 'pageprops' in QueryGenerator
Currently, 'pageprops' is only cached if found in the response.
The API only returns this property if it's not empty, causing
repetitive queries if accessed multiple times.
This change caches 'pageprops' if it's requested, creating an
empty dictionary to use if it's not present in the response.

Patch c0b61ad already does this for other properties such as
templates and langlinks.

Change-Id: I30973b7479822ebc043f1d4db77522143e8c5652