Break import cycle between page and cc modules
pywikibot.cosmetic_changes imported pagegenerators only for
parameterHelp, which is unused since the module was moved into
the pywikibot package in fd1ba1e48e.

Also move the import of page in cosmetic_changes down to
where it is used, so that cosmetic_changes be a normal
import in the page module.

Change-Id: I5ec082c65247129454b605781210f09c168b4a51