[WIP] Fix ItemPage redirects
When a item is a redirect, wbgetentities by default returns
a `redirects` mapping and the data for the target.  The
`redirects` mapping can be used to determine the redirect
status and the redirect target.
This can be disabled by using the parameter `redirects`, but
that was a recent addition, and then it replies that the
redirect is a missing item.

WikibasePage classes get() method then populates attributes like
`labels`, `claims` and `sitelinks` with the values of the target;
these need to be cleared, and can be used to preload the target
if get's get_redirect method was enabled.

exists() should determine the status, even if it is a redirect.

editEntity and toJSON should raise NotImplementedError for redirects.

Change-Id: I9d49992cb27dd30faabd31328be9bb28f96d8c55