[FEAT] Implement read only ProxyDict
This is adding a read only implementation of a dict mapping to another dict. So
changes in the underlying dict are reflected by the `ProxyDict`. This can be
used for example for classes that only provide read only mapping capabilities.

This class can be used instead of `MappingProxyType` in Python 3 it it should
be subclassed. And this also provides a `MappingProxyType` in Python 2 similar
to the type in Python 3. The differences are that it's subclassable and has a
`create` classmethod.

Change-Id: I25845b239c7d200168679b7cc737257502985940