[IMPROV] listpages: Create Formatter once
Instead of creating a `Formatter` instance on every page it's only creating
once, defining the format string and verifying that the output language is set.

It also makes `num` in `output` mandatory as most default format strings do not
work without defining it and the script is defining it always anyways. And if a
script is not needing it, it doesn't matter that it's supplied. The preset
format strings can now be references by an integer. The default value of
`output` was an int and wouldn't have worked anyway if the script hadn't
supplied it.

This is also using `color_format` so it's no longer necessary to prefix colors
using `\03` and escaping them in double curly brackets.

Change-Id: I2b145cb6cbc45ddb8b3dae38397302d0da8c7842