[IMPROV] Expect failure instead of skipping it
The `testWebCiteOlder` test in `tests.weblib_tests.TestWebCite` doesn't work
because WebCite implementation doesn't work anymore. The memento implementation
though does work so it shouldn't expect a failure there.

To avoid this, 544f3207 skipped that test but it could still be executed
manually and fail then. Expecting a failure it won't fail but require that.

Change-Id: I065f8ce0a63e710d22129d0f7632d6e3f4a46ed8