[FEAT] cosmetic_changes: Replace helper method
Instead of using `replaceExcept` everywhere and to handle `text` manually this
adds a `replace` method which edits an attribute and thus it's not necessary
for the caller to provide the text or store its result. It also uses a sensible
set of default exceptions which can be extended or overwritten in case they
don't actually match the need.

Change-Id: I9f570f4d424ab6a5d6fdcb17268cdf4e164838ea