Return numbers in WbQuantity-json-object as strings with sign
The 'to Wikibase()' function currently returns a dictionary
where the values for the keys 'amount', 'upperBound' and
'lowerBound' are returned as int or float. This caused the
numbers to be deserialized into very long floats
(>50 digits) on Wikidata. These long numbers make the diffs
hard to read, and the values hard to read and edit.

The patch formats the numbers as strings that are prepended
with the correct sign. 14.1 becomes '+14.1' and -14.1
becomes '-14.1'. The tests were modified to work with these
changes.

Change-Id: Ib54be2898a035a99592cba1298f354bd1637a6b5