Add tests for Claim fromJSON-methods
This patch adds tests to 'page_tests.py' for the Claim class in
'page.py'. The test creates 3 Claim objects using the 'fromJSON',
'referenceFromJSON' and 'qualifierFromJSON' methods. The
objects are created from dictionaries. For each object the tests
make sure that 'claim_obj.isReference' and 'claim_obj.isQualifier'
have the correct True and False value.

The tests were written for 244834 and 239780

Change-Id: I12b4fc688cc38380fef0b20ac44b1a129b649892