[IMPROV] Simplify findmarker usage
In most cases the marker got by `findmarker` consists of one character and it
starts using that twice and adds that one character each time. And in theory
any other way is not necessary to determine a marker. So this is replacing
`findmarker` by `generate_marker` which only accepts one character as a
parameter.

Change-Id: I5c253dffbf61ddc205eb4429c60aab9314c149ef