OAuth: support Site login with (username, token)
OAuth token can now be provided when creating Site instance, in order to
use OAuth without necessarily store tokens in config files.

This is useful when using pywikibot as a library to login a user at run
time, once oauth token data have been fetched.

The selection of which data to associate to a Site instance is now
done in Site factory in __init__.py.

If a user and its OAuth tokens are passed to the Site factory, a Site
instance supporting login of such user is created.
Otherwise it creates a Site object from data in config file (including
Oauth data in config.authenticate if required).

Get OAuth configuration and login via Oauth in LoginManager via th new
function LoginManager.login_oauth().

Deprecate and rename LoginManager.getCookie() to
LoginManager.login_to_site, as the getCookie() name is misleading.

Remove current implemention of LoginManager.storecookiedata() and raise
NotImplementedError, as the real handling of cookies is done in the
subclass api.LoginManager.

Requests are done in http.py fetching the auth data from the site
object in the request.

Bug: T118012
Change-Id: If192745e05bf77920e0fe6a5f3b4361077322734