[Cleanup] Deprecate special_page_limit.
special_page_limit was introduced in compat for retrieving pages from
mw special list via screen scraping. The use of this parameter in compat
is a mess but in core it is not used except of one method descripbed below.
api may read all items except a total parameter for a generator or treat
counter restricts it.

The only use of that parameter is inside Site.watched_page()
but it doesn't make any sense to have that default value for
the total amount of retrieved items.

Change-Id: Idd0afa55cec93b92abbd5f0a9d2d0b9317a9a42a