[IMPR] Provide preloading via GeneratorFactory
Also disable PreloadingGenerator when -liverecentchanges is selected

It is exceedingly superfluous to make a bulk load for -liverecentchanges which
waits for an event. Otherwise the process waits until the buffer is full to
start the operation. This is counterproductive because that option intends
processing the event as soon as possible.

Bug: T135331
Change-Id: I52777a8b2405686ad7a3dffdf631789752afb7ff