Support adding units to WbQuantity through the entity url
This add the entity parameter to WbQuantity as a way of supplying
it with a unit.

entity/_entity is implemented in a similar way to (globe) entity
in Coordinate so that it can still be used even when units get
supported through the unit parameter.

Note that units need not be items in the same repo as the claim,
as sucht this mechanism is needed even after full unit support is
implemented.

This is part 2 of 3 of the commits to allow unit ItemPages
to be passed to WbQuantity.

Bug: T143594
Change-Id: Ibc05165841df8c1ef3108a55dc53bfcbe09da35d