Support adding units to WbQuantity through ItemPage
This is be backwards compatible but will raise a warning for
the users of the undocumented feature of providing the url
to the unit parameter.

Note that the representation of WbQuantity changes somewhat
(unit='1' -> unit=None), again this is the same difference
as in Coordinate.

This is part 3 of 3 of the commits to allow unit ItemPages
to be passed to WbQuantity.

Bug: T143594

Change-Id: Icdcfed92df7734d5162c432a2c6d3d00a1526491