Make sparql use Site to access sparql endpoint and entity_url
Introduces a repo parameter which takes any DataSite and gets
its sparql endpoint and entity_url. Still supports explicit
endpoint and entity_url for full flexibility.

As before this defaults to Wikidata but now relies on its Site
object rather than hard-coded endpoint and entity_url.

The same change is implemented in pagegenerators.

Note that this change is breaking for anyone who had overridden
endpoint but not entity_url (i.e. running their own endpoint but
loaded with wikidata data). This was an undocumented feature and
these users will now recieve an Error.

Change-Id: I057335a56398a7d5f6c8c811233a002228b73093