[bugfix] Manage temporary readonly error
It's look like the API error named internal_api_error_ReadOnlyError was renamed to readonly. So, code.startswith(u'internal_api_error_') is not enabled. I've edited the code to accept readonly as an error that need edit retry. Need investigation to verify if internal_api_error_ReadOnlyError is still enabled, and if others error codes are impacted.

Bug: T154011
Change-Id: I9cc80c77532db3d9b71db5c33f91b8a699f41d6c