Allow retrieval of unit as ItemPage for WbQuantity
Introduces ItemPage.from_entity_url() which attempts to retrieve
the ItemPage corresponding to an entity url
(e.g. http://www.wikidata.org/entity/Q123)

This function is accessed from WbQuantity.get_unit_item()
to retrieve the ItemPage correspoding to the unit.

In future it could also be used for e.g. Coordinate globes or
to resolve sparql results.

The site parameter was added to TestWbQuantity tests so as to
obey rules for Dry tests.

Bug: T143594
Change-Id: I9746881ada22cb7dd47a76c0597737f063cd31f6