Encode default globe in family file.
The current default assumes that every Wikibase installation has a
globe named 'earth' which need not be the case. It also assigns
this globe even if overridden by globe_item.

Instead encode the default globe in the family file and only fall
back on the default if neither globe nor globe_item has been set.
Also clarifies that globe_item takes precedence by mentioning this
in docstring and using entity instead of globe in representation.

Note that 'earth' is still implicitly implied in
Coordinate.precision.

Change-Id: I1a13153ebd62e6eb4efbe05cbe09f0bac748ed99