Port rcstream to EventStreams recentchange stream
RCStream will be decomissioned in July.  This change moves the core
streaming functionality to evenstreams.py.  rcstream.py
classes and functions have the same interfaces they had before, but
now they wrap eventstreams.  This should be a backwards
compatible change, and should allow for future EventStreams
endpoints to be added as well.

Bug: T158943
Change-Id: Ie21d19cdb452122420885617b3270e9aecebc05b