Use global option -code instead of -lang
Also change option -tolang to -tocode in transferbot.py

site.code is the identifying code of a site presumed to be equal to the
wiki prefix whereas site language is the ISO language code for a site.

e.g. 'als' is a site code whereas its site language is 'gsw'
     'meta' is a site code whereas its language is 'en'

The global option should follow this difference.

Change-Id: I145c0dc283ff597a697752aec901c1a14e6b1416