Remove TODO remark in pwb.py
pwb.py is a wrapper script to enable the pywikibot framework if it is not
installed as a side package. It should not use and it should not have to use
the package itself. We have the same issue with the print function.
We don't use the pywikibot.output() inside and we never should do it.

Change-Id: I4edfef43b538c411847eafa2ee279d632b0bbd61