[IMPR] Make socket_timeout reusable
On older installations of requests it's not possible to use two timeout values
but newer versions may have two: one for the connection one for read timeout.

To ensure the right value depending on the requests version there as a fix
inside http.py; this should be moved to the upper library place to be
reusable by other library scripts like rcstreams or eventstreams.

__init__.py:
- Add a new function method to return the socket_timeout dependig on
  requests version

- http.py:
  remove the socket_timeout fixer and use default_session_timeout() instead.

- Since requests > 2.4.1 is required with T134647 for pwb now
  the default socket_timeout may be a tuple again.

Bug: T166539
Change-Id: I598ac9fa2fa4f237b68d44bd8cf3c587d5768bd1