site_tests.testLinkMethods: Ignore errors caused by page content change
If page content is changed during the test, set of its templates or
wiki-links may also change and cause an assertion error.

In such situations, assert that the links/templates have actually changed,
update the cache (for the next run), and ignore the error.

Use unittest.mock to change the returned value of TestRequest._expired to
True. Otherwise it will always return False and will not allow to update
the cache.

Remove the unittest_print that was previously added for debugging.

Bug: T92856
Bug: T153501
Change-Id: I3b243929a46fafc000b8f175cfa99d32b697941d